+++ Library/digitalWrite.h
@@ -2,5 +2,5 @@
 #include <avr/io.h>
 #include "avr_uno.h"
 
-short int digitalWrite(short int apin, short int level) ;
+uint8_t digitalWrite(uint8_t apin, uint8_t level) ;
