+++ Library/pinMode.h
@@ -2,7 +2,7 @@
 #include <avr/io.h>
 #include "avr_uno.h"
 
-short int pinMode(short int apin, short int mode) ;
+uint8_t pinMode(uint8_t apin, uint8_t mode) ;
 
-short int digitalWrite(short int apin, short int level) ;
+uint8_t digitalWrite(uint8_t apin, uint8_t level) ;
