+++ examples/pinMode/pinMode.c
@@ -6,7 +6,7 @@
 
 int main (void)
 {
-    short int UNO_pin = 13;
+    uint8_t UNO_pin = 13;
     on_led();
     pinMode(UNO_pin, OUTPUT);
     for (int i=0;i<40;i++) {
